light_poseidon::parameters

Module bn254_x5

Constants and MDS matrix for the BN254 curve with the following properties:

x^5 S-boxes

3 prime fields (one zero prime field and two inputs from the caller)

8 full rounds and 57 partial rounds

Those parameters are used for our Poseidon hash implementation.

They were generated using the official script from the Poseidon paper: generate_parameters_grain.sage with the following parameters:

sage generate_parameters_grain.sage 1 0 254 3 8 57 0x30644e72e131a029b85045b68181585d2833e84879b9709143e1f593f0000001

Constants

ALPHA

FULL_ROUNDS

PARTIAL_ROUNDS

Functions

get_poseidon_parameters